+++ Sources/Sentry/Public/SentryOptions.h
@@ -13,8 +13,10 @@ NS_ASSUME_NONNULL_BEGIN
 @class SentryHttpStatusCodeRange;
 @class SentryMeasurementValue;
 @class SentryReplayOptions;
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+@class SentryProfileOptions;
+#endif // SENTRY_TARGET_PROFILING_SUPPORTED
 @class SentryScope;
-@class SentryReplayOptions;
 
 NS_SWIFT_NAME(Options)
 @interface SentryOptions : NSObject
@@ -515,6 +517,20 @@ NS_SWIFT_NAME(Options)
 
 #if SENTRY_TARGET_PROFILING_SUPPORTED
 /**
+ * Block used to configure the continuous profiling options.
+ * @warning Continuous profiling is an experimental feature and may contain bugs.
+ * @seealso @c SentryProfileOptions, @c SentrySDK.startProfiler and @c SentrySDK.stopProfiler .
+ */
+typedef void (^SentryProfilingConfigurationBlock)(SentryProfileOptions *_Nonnull profiling);
+
+/**
+ * Configuration for the Sentry profiler.
+ * @warning: Continuous profiling is an experimental feature and may still contain bugs.
+ * @warning: Profiling is automatically disabled if a thread sanitizer is attached.
+ */
+@property (nullable, nonatomic, copy) SentryProfilingConfigurationBlock configureProfiling;
+
+/**
  * @warning This is an experimental feature and may still have bugs.
  * Set to @c YES to run the profiler as early as possible in an app launch, before you would
  * normally have the opportunity to call @c SentrySDK.start . If @c profilesSampleRate is nonnull,
@@ -523,6 +539,8 @@ NS_SWIFT_NAME(Options)
  * @warning If @c profilesSampleRate is @c nil then a continuous profile will be started on every
  * launch; if you desire sampling profiled launches, you must compute your own sample rate to decide
  * whether to set this property to @c YES or @c NO .
+ * @warning This property is deprecated and will be removed in a future version of the SDK. See
+ * @c SentryProfileOptions.startOnAppStart and @c SentryProfileOptions.lifecycle .
  * @note Profiling is automatically disabled if a thread sanitizer is attached.
  */
 @property (nonatomic, assign) BOOL enableAppLaunchProfiling;
@@ -549,6 +567,8 @@ NS_SWIFT_NAME(Options)
  * @note The default is @c nil (which implies continuous profiling mode).
  * @warning The new continuous profiling mode is experimental and may still contain bugs.
  * @note Profiling is automatically disabled if a thread sanitizer is attached.
+ * @warning This property is deprecated and will be removed in a future version of the SDK. See
+ * @c  SentryProfileOptions.sessionSampleRate.
  */
 @property (nullable, nonatomic, strong) NSNumber *profilesSampleRate;
 
@@ -561,6 +581,8 @@ NS_SWIFT_NAME(Options)
  * with @c SentrySamplingContext.forNextAppLaunch set to @c YES, and the result will be persisted to
  * disk for use on the next app launch.
  * @note Profiling is automatically disabled if a thread sanitizer is attached.
+ * @warning This property is deprecated and will be removed in a future version of the SDK. See
+ * @c SentryProfileOptions.sessionSampleRate .
  */
 @property (nullable, nonatomic) SentryTracesSamplerCallback profilesSampler NS_SWIFT_SENDABLE;
 
@@ -573,6 +595,7 @@ NS_SWIFT_NAME(Options)
  * @returns @c YES if either @c profilesSampleRate > @c 0 and \<= @c 1 , or @c profilesSampler is
  * set, otherwise @c NO.
  * @note Profiling is automatically disabled if a thread sanitizer is attached.
+ * @warning This property is deprecated and will be removed in a future version of the SDK.
  */
 @property (nonatomic, assign, readonly) BOOL isProfilingEnabled;
 
@@ -623,8 +646,6 @@ NS_SWIFT_NAME(Options)
  * @note This flag wins over enableAppHangTracking. When enabling both enableAppHangTracking and
  enableAppHangTrackingV2, the SDK only enables enableAppHangTrackingV2 and disables
  enableAppHangTracking.
- *
- * @warning This is an experimental feature and may still have bugs.
  */
 @property (nonatomic, assign) BOOL enableAppHangTrackingV2;