+++ Sources/Sentry/SentryProfiler.mm
@@ -5,8 +5,10 @@
 #    import "SentryContinuousProfiler.h"
 #    import "SentryDependencyContainer.h"
 #    import "SentryDispatchQueueWrapper.h"
+#    import "SentryFileManager.h"
 #    import "SentryFramesTracker.h"
 #    import "SentryHub+Private.h"
+#    import "SentryInternalDefines.h"
 #    import "SentryLaunchProfiling.h"
 #    import "SentryLog.h"
 #    import "SentryMetricProfiler.h"
@@ -14,6 +16,7 @@
 #    import "SentryProfilerState+ObjCpp.h"
 #    import "SentryProfilerTestHelpers.h"
 #    import "SentrySDK+Private.h"
+#    import "SentrySampling.h"
 #    import "SentrySamplingProfiler.hpp"
 #    import "SentryScreenFrames.h"
 #    import "SentrySwift.h"
@@ -28,6 +31,8 @@
 
 using namespace sentry::profiling;
 
+SentrySamplerDecision *_Nullable sentry_profilerSessionSampleDecision;
+
 namespace {
 
 static const int kSentryProfilerFrequencyHz = 101;
@@ -37,20 +42,94 @@ static const int kSentryProfilerFrequencyHz = 101;
 #    pragma mark - Public
 
 void
-sentry_manageTraceProfilerOnStartSDK(SentryOptions *options, SentryHub *hub)
+sentry_reevaluateSessionSampleRate(float sessionSampleRate)
 {
+    sentry_profilerSessionSampleDecision = sentry_sampleProfileSession(sessionSampleRate);
+}
+
+void
+sentry_configureContinuousProfiling(SentryOptions *options)
+{
+    if (![options isContinuousProfilingEnabled]) {
+        if (options.configureProfiling != nil) {
+            SENTRY_LOG_WARN(@"In order to configure SentryProfileOptions you must remove "
+                            @"configuration of the older SentryOptions.profilesSampleRate, "
+                            @"SentryOptions.profilesSampler and/or SentryOptions.enableProfiling");
+        }
+        return;
+    }
+
+    if (options.configureProfiling == nil) {
+        SENTRY_LOG_DEBUG(@"Continuous profiling V2 configuration not set by SDK consumer, nothing "
+                         @"to do here.");
+        return;
+    }
+
+    options.profiling = [[SentryProfileOptions alloc] init];
+    options.configureProfiling(options.profiling);
+
+    if (options.profiling.lifecycle == SentryProfileLifecycleTrace && !options.isTracingEnabled) {
+        SENTRY_LOG_WARN(
+            @"Tracing must be enabled in order to configure profiling with trace lifecycle.");
+        return;
+    }
+
+    sentry_reevaluateSessionSampleRate(options.profiling.sessionSampleRate);
+
+    SENTRY_LOG_DEBUG(@"Configured profiling options: <%@: {\n  lifecycle: %@\n  sessionSampleRate: "
+                     @"%.2f\n  profileAppStarts: %@\n}",
+        options.profiling,
+        options.profiling.lifecycle == SentryProfileLifecycleTrace ? @"trace" : @"manual",
+        options.profiling.sessionSampleRate, options.profiling.profileAppStarts ? @"YES" : @"NO");
+}
+
+void
+sentry_sdkInitProfilerTasks(SentryOptions *options, SentryHub *hub)
+{
+    sentry_configureContinuousProfiling(options);
+
     [SentryDependencyContainer.sharedInstance.dispatchQueueWrapper dispatchAsyncWithBlock:^{
-        BOOL shouldStopAndTransmitLaunchProfile = options.profilesSampleRate != nil;
+        // get the configuration options from the last time the launch config was written; it may be
+        // different than the new options the SDK was just started with
+        const auto configDict = sentry_appLaunchProfileConfiguration();
+        const auto profileIsContinuousV1 =
+            [configDict[kSentryLaunchProfileConfigKeyContinuousProfiling] boolValue];
+        const auto profileIsContinuousV2 =
+            [configDict[kSentryLaunchProfileConfigKeyContinuousProfilingV2] boolValue];
+        const auto v2LifecycleValue
+            = configDict[kSentryLaunchProfileConfigKeyContinuousProfilingV2Lifecycle];
+        const auto v2Lifecycle = (SentryProfileLifecycle)
+            [configDict[kSentryLaunchProfileConfigKeyContinuousProfilingV2Lifecycle] intValue];
+        const auto v2LifecycleIsManual = profileIsContinuousV2 && v2LifecycleValue != nil
+            && v2Lifecycle == SentryProfileLifecycleManual;
+
+        BOOL shouldStopAndTransmitLaunchProfile = YES;
+
 #    if SENTRY_HAS_UIKIT
-        if (SentryUIViewControllerPerformanceTracker.shared.alwaysWaitForFullDisplay) {
+        const auto v2LifecycleIsTrace = profileIsContinuousV2 && v2LifecycleValue != nil
+            && v2Lifecycle == SentryProfileLifecycleTrace;
+        const auto profileIsCorrelatedToTrace = !profileIsContinuousV2 || v2LifecycleIsTrace;
+        if (profileIsCorrelatedToTrace
+            && SentryUIViewControllerPerformanceTracker.shared.alwaysWaitForFullDisplay) {
+            SENTRY_LOG_DEBUG(@"Will wait to stop launch profile correlated to a trace until full "
+                             @"display reported.");
             shouldStopAndTransmitLaunchProfile = NO;
         }
 #    endif // SENTRY_HAS_UIKIT
+
+        if (profileIsContinuousV1 || v2LifecycleIsManual) {
+            SENTRY_LOG_DEBUG(
+                @"Continuous manual launch profiles aren't stopped on calls to SentrySDK.start, "
+                @"not stopping profile.");
+            shouldStopAndTransmitLaunchProfile = NO;
+        }
+
         if (shouldStopAndTransmitLaunchProfile) {
             SENTRY_LOG_DEBUG(@"Stopping launch profile in SentrySDK.start because there will "
                              @"be no automatic trace to attach it to.");
             sentry_stopAndTransmitLaunchProfile(hub);
         }
+
         sentry_configureLaunchProfiling(options);
     }];
 }
@@ -65,6 +144,14 @@ sentry_manageTraceProfilerOnStartSDK(SentryOptions *options, SentryHub *hub)
     // we want to allow starting a launch profile from here for UI tests, but not unit tests
     if (NSProcessInfo.processInfo.environment[@"--io.sentry.ui-test.test-name"] == nil) {
         return;
+    }
+
+    // the ui tests want to wipe the data before each test case runs, to remove any launch config
+    // files that might be present before launching the app initially, however we need to make sure
+    // to remove stale versions of the file before it gets used to potentially start a launch
+    // profile that shouldn't have started, so we check here for this
+    if ([NSProcessInfo.processInfo.arguments containsObject:@"--io.sentry.wipe-data"]) {
+        removeSentryStaticBasePath();
     }
 #    endif // defined(SENTRY_TEST) || defined(SENTRY_TEST_CI)
     sentry_startLaunchProfile();