+++ Sources/Sentry/include/SentryFileManager.h
@@ -120,7 +120,8 @@ BOOL createDirectoryIfNotExists(NSString *path, NSError **error);
 /**
  * Path for a default directory Sentry can use in the app sandbox' caches directory.
  * @note This method must be statically accessible because it will be called during app launch,
- * before any instance of @c SentryFileManager exists, and so wouldn't be able to access this path
+ * before any instance of @c SentryFileManager exists, and so wouldn't be able to access this path.
+ * @note For unsandboxed macOS apps, the path has the form @c ~/Library/Caches/<app-bundle-id> .
  * from an objc property on it like the other paths. It also cannot use
  * @c SentryOptions.cacheDirectoryPath since this can be called before
  * @c SentrySDK.startWithOptions .
@@ -143,7 +144,8 @@ SENTRY_EXTERN BOOL appLaunchProfileConfigFileExists(void);
  * Retrieve the contents of the launch profile config file, which stores the sample rates used to
  * decide whether or not to profile this launch.
  */
-SENTRY_EXTERN NSDictionary<NSString *, NSNumber *> *_Nullable appLaunchProfileConfiguration(void);
+SENTRY_EXTERN NSDictionary<NSString *, NSNumber *> *_Nullable sentry_appLaunchProfileConfiguration(
+    void);
 
 /**
  * Write a config file that stores the sample rates used to determine whether this launch should
@@ -159,6 +161,12 @@ SENTRY_EXTERN void writeAppLaunchProfilingConfigFile(
  * start the profiler.
  */
 SENTRY_EXTERN void removeAppLaunchProfilingConfigFile(void);
+
+SENTRY_EXTERN NSString *_Nullable sentryStaticBasePath(void);
+
+#    if defined(SENTRY_TEST) || defined(SENTRY_TEST_CI)
+SENTRY_EXTERN void removeSentryStaticBasePath(void);
+#    endif // defined(SENTRY_TEST) || defined(SENTRY_TEST_CI)
 
 #endif // SENTRY_TARGET_PROFILING_SUPPORTED