+++ Sources/Swift/Helper/SentryEnabledFeaturesBuilder.swift
@@ -2,7 +2,7 @@ import Foundation
 
 @objcMembers class SentryEnabledFeaturesBuilder: NSObject {
 
-    // swiftlint:disable cyclomatic_complexity
+    // swiftlint:disable cyclomatic_complexity function_body_length
     static func getEnabledFeatures(options: Options?) -> [String] {
         guard let options = options else {
             return []
@@ -49,15 +49,23 @@ import Foundation
             features.append("persistingTracesWhenCrashing")
         }
 
-#if os(iOS) && !SENTRY_NO_UIKIT
-        if options.sessionReplay.enableExperimentalViewRenderer {
+#if (os(iOS) || os(tvOS)) && !SENTRY_NO_UIKIT
+        if options.sessionReplay.enableViewRendererV2 {
+            // We keep the old name for backwards compatibility of the telemetry data.
             features.append("experimentalViewRenderer")
         }
         if options.sessionReplay.enableFastViewRendering {
             features.append("fastViewRendering")
         }
-#endif // #if os(iOS) && !SENTRY_NO_UIKIT
+#endif // (os(iOS) || os(tvOS)) && !SENTRY_NO_UIKIT
+
+        if options.experimental.enableDataSwizzling {
+            features.append("dataSwizzling")
+        }
+        if options.experimental.enableFileManagerSwizzling {
+            features.append("fileManagerSwizzling")
+        }
         return features
     }
-    // swiftlint:enable cyclomatic_complexity
+    // swiftlint:enable cyclomatic_complexity function_body_length
 }